restructure src/view folder and its vue file content by adding intermediate folder
