+++ src/router/index.js
@@ -3,7 +3,7 @@ import VueRouter from 'vue-router';
 import Home from '../views/Home.vue';
 import SiteSearch from '../views/SiteSearch.vue';
 import Board from '../views/Board.vue';
-import ListOfPartners from '../views/ListOfPartners.vue';
+import ListOfPartners from '../views/PartnerInformation/ListOfPartners.vue';
 import CNA from '../views/CNA.vue';
 
 Vue.use(VueRouter);
