+++ src/views/PartnerInformation/ListOfPartners.vue
@@ -177,7 +177,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import '../assets/style/globals.scss';
+@import '../../assets/style/globals.scss';
 
 .cve-main-container {
     text-align: left;
